+++ my_iot/services/clock.py
@@ -8,12 +8,12 @@ from typing import Optional
 from loguru import logger
 
 from my_iot.services.base import Service
-from my_iot.types_ import Event
+from my_iot.types_ import Event, Unit
 
 
 class Clock(Service):
-    def __init__(self, channel_id: str, interval: timedelta, title: Optional[str] = None):
-        self.channel_id = channel_id
+    def __init__(self, sub_channel: str, interval: timedelta, title: Optional[str] = None):
+        self.sub_channel = sub_channel
         self.interval = interval.total_seconds()
         self.title = title
 
@@ -22,8 +22,7 @@ class Clock(Service):
         for i in count(start=1):
             logger.trace('Sleeping for {interval} seconds…', interval=self.interval)
             await sleep(self.interval)
-            # FIXME: set unit.
-            yield Event(channel_id=f'clock:{self.channel_id}', value=i, title=self.title)
+            yield Event(channel=f'clock:{self.sub_channel}', value=i, title=self.title, unit=Unit.TEXT)
 
     def __str__(self) -> str:
-        return f'Clock(key={self.channel_id!r}, interval={self.interval!r}, title={self.title!r})'
+        return f'Clock(key={self.sub_channel!r}, interval={self.interval!r}, title={self.title!r})'